# PERF: Optimize PasswordValidatorService.cs