--- a/src/SimplePasswordCheck/SimplePasswordCheck.Benchmark/Password/Service/PasswordValidatorService_Benchmark.cs
+++ b/src/SimplePasswordCheck/SimplePasswordCheck.Benchmark/Password/Service/PasswordValidatorService_Benchmark.cs
@@ -8,6 +8,7 @@ using System.Threading.Tasks;
 
 namespace SimplePasswordCheck.Benchmark.Password.Service
 {  
+    [MemoryDiagnoser]
     public class PasswordValidatorServiceBenchmark
     {
         [Params("AbTp9!fok",